Update all objects.h, rename counts -> top_count
