+++ targets/TARGET_STM/TARGET_STM32F7/objects.h
@@ -65,16 +65,6 @@ struct trng_s {
     RNG_HandleTypeDef handle;
 };
 
-struct pwmout_s {
-    PWMName pwm;
-    PinName pin;
-    uint32_t prescaler;
-    uint32_t period;
-    uint32_t pulse;
-    uint8_t channel;
-    uint8_t inverted;
-};
-
 struct serial_s {
     UARTName uart;
     int index; // Used by irq
