+++ targets/TARGET_STM/TARGET_STM32L4/objects.h
@@ -44,16 +44,6 @@ struct dac_s {
 };
 #endif
 
-struct pwmout_s {
-    PWMName pwm;
-    PinName pin;
-    uint32_t prescaler;
-    uint32_t period;
-    uint32_t pulse;
-    uint8_t channel;
-    uint8_t inverted;
-};
-
 struct serial_s {
     UARTName uart;
     int index; // Used by irq
