+++ targets/TARGET_STM/TARGET_STM32WB/objects.h
@@ -37,16 +37,6 @@
 extern "C" {
 #endif
 
-struct pwmout_s {
-    PWMName pwm;
-    PinName pin;
-    uint32_t prescaler;
-    uint32_t period;
-    uint32_t pulse;
-    uint8_t channel;
-    uint8_t inverted;
-};
-
 struct serial_s {
     UARTName uart;
     int index; // Used by irq
