+++ targets/TARGET_STM/device.h
@@ -38,6 +38,7 @@
 #include "objects.h"
 #include "stm_i2c_api.h"
 #include "stm_spi_api.h"
+#include "stm_pwmout_api.h"
 
 #if DEVICE_USTICKER
 #include "us_ticker_defines.h"
